refactor recharge function
