+++ server/database.js
@@ -4,9 +4,9 @@ module.exports = {
     datebaseConnection(dbURL) {
         setTimeout(function () {
             mongoose
-                .connect(dbURL, { useNewUrlParser: true })
+                .createConnection(dbURL, { useNewUrlParser: true })
                 .then(x => {
-                    console.log(`Connected to Mongo! Database name: "${x.connections[0].name}"`)
+                    console.log(`Connected to Mongo! Database name: "${x.name}"`)
                 })
                 .catch(err => {
                     console.error('Error connecting to mongo', err)
@@ -14,4 +14,3 @@ module.exports = {
         }, 10000)
     }
 }
-
