+++ server/services/validation/validationCredit.js
@@ -1,10 +1,7 @@
-const Credit = require('../../models/Credit')
-var locks = require('locks');
-var mutex = locks.createMutex();
-
+var rechargeCredit = require('../credit/rechargeCredit')
 
 module.exports = {
-    recharge(id, amountCharge, res) {
+    validate(id, amountCharge, res) {
         if (amountCharge == '') {
             res.status(400).json("Amount is missing")
         }
@@ -16,21 +13,7 @@ module.exports = {
         } else if (amountCharge < 1 || amountCharge > 500) {
             res.status(400).json("Amount needs to be between 1 and 500")
         } else {
-            mutex.lock(function () {
-                Credit.findByIdAndUpdate({ _id: id }, {
-                    $inc: {
-                        amount: amountCharge
-                    }
-                }, { new: true })
-                    .then(() => {
-                        mutex.unlock()
-                        res.status(200).json('Credit re-established')
-                    })
-                    .catch(err => {
-                        mutex.unlock()
-                        res.status(200).json(err)
-                    })
-            })
+            rechargeCredit.recharge(id, amountCharge, res)
         }
     }
 }
